+++ src/main/java/event/control/FinSimulation.java
@@ -6,23 +6,21 @@ import main.java.Controller;
 import main.java.Event;
 import main.java.Scheduler;
 
-import javax.naming.ldap.Control;
-
 public class FinSimulation extends Event {
     /**
      * Instantiates a new main.java.Event.
      *
      * @param startingTime the starting time
      */
-    public FinSimulation(float startingTime) {
+    public FinSimulation(double startingTime) {
         super(startingTime);
     }
 
     @Override
     public void run() {
         Scheduler.getInstance().cleanScheduler();
-        float meanWaitingTimeBeforeControl = Controller.getInstance().getControlQueueArea()/ Controller.getInstance().getNbBusEntres();
-        float meanWaitingTimeBeforeReparation = Controller.getInstance().getRepairQueueArea()/Controller.getInstance().getNbBusRepares();
-        float ReparationCenterUseRate =Controller.getInstance().getOccupationArea()/(2*160);
+        double meanWaitingTimeBeforeControl = Controller.getInstance().getControlQueueArea()/ Controller.getInstance().getNbBusEntres();
+        double meanWaitingTimeBeforeReparation = Controller.getInstance().getRepairQueueArea()/Controller.getInstance().getNbBusRepares();
+        double ReparationCenterUseRate =Controller.getInstance().getOccupationArea()/(2*160);
     }
 }
